fix(tui): refine plugin update states
